+++ api/src/main/java/org/allaymc/api/container/RecipeContainer.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.container;
+
+import org.allaymc.api.item.recipe.input.RecipeInput;
+
+/**
+ * @author IWareQ
+ */
+public interface RecipeContainer {
+    RecipeInput createRecipeInput();
+}