+++ server/src/main/java/org/allaymc/server/blockentity/component/BlockEntityBrewingStandBaseComponentImpl.java
@@ -60,6 +60,7 @@ public class BlockEntityBrewingStandBaseComponentImpl extends BlockEntityBaseCom
             container.addOnSlotChangeListener(finalI, item -> {
                 var currentBlockState = getDimension().getBlockState(position);
                 var values = currentBlockState.getBlockType().getProperties().values();
+                //noinspection unchecked
                 currentBlockState = currentBlockState.setPropertyValue(values.toArray(new BlockPropertyType[0])[finalI - 1], item != ItemAirStack.AIR_STACK);
                 getDimension().setBlockState(
                         position, currentBlockState,